requirements.txt
fastapi==0.115.9
uvicorn==0.34.0

numpy==2.2.3
pandas==2.2.3
sshtunnel == 0.4.0
python-dotenv==1.0.1
pymysql==1.1.1
